Allow specifiers to be partially defined in dictionaries
With this update the feature developed over the last 3 commits is done.

It allows to specify at the module level new specifier without touching
the logic of the code. This allows the definition of a new variation of
the patterns to implement a new side effect.

The new side effect this was designed for is the "not_item" version.
Let's things not be itemized in latex, but still counted when generating
the counts, and still appear in the correct color.